Split the I2C busses and added the RPi-SWD connection.
It was found that when the RPi was powered off, the I2C connection was lost.
This was because the RPI has pull-ups on board that would act as voltage dividers.

The SWD connection to the RPi was made via jumpers to allow programming and
debugging of the STM directly via the Pi.